upgrade walletconnect to reown
