+++ package.json
@@ -4,6 +4,9 @@
   "private": true,
   "type": "module",
   "dependencies": {
+    "@reown/appkit": "^1.7.4",
+    "@reown/appkit-adapter-wagmi": "^1.7.4",
+    "@tanstack/react-query": "^5.76.0",
     "axios": "^1.7.2"
   },
   "scripts": {
@@ -27,13 +30,19 @@
   },
   "devDependencies": {
     "@apollo/client": "^3.11.4",
+    "@ethereumjs/util": "^9.0.3",
     "@graphql-codegen/cli": "^5.0.3",
     "@graphql-codegen/typescript": "^4.1.0",
     "@graphql-codegen/typescript-operations": "^4.3.0",
+    "@metamask/eth-sig-util": "^7.0.2",
     "@playwright/test": "^1.44.1",
+    "@sentry/sveltekit": "^8",
     "@sveltejs/adapter-auto": "^3.2.1",
+    "@sveltejs/adapter-node": "^5.2.5",
     "@sveltejs/adapter-static": "^3.0.1",
+    "@sveltejs/adapter-vercel": "^5.3.1",
     "@sveltejs/kit": "^2.5.10",
+    "@sveltejs/vite-plugin-svelte": "^3.1.1",
     "@types/debug": "^4.1.12",
     "@types/node": "^22.10.1",
     "@types/object-hash": "^3.0.6",
@@ -42,68 +51,62 @@
     "@vercel/analytics": "^1.3.1",
     "@vitest/coverage-v8": "^1.6.0",
     "@wagmi/cli": "^2.1.8",
+    "@wagmi/connectors": "^5.0.9",
+    "@wagmi/core": "^2.16.1",
+    "@walletconnect/core": "^2.17.3",
+    "@walletconnect/utils": "^2.17.3",
+    "@walletconnect/web3wallet": "^1.16.1",
+    "@web3modal/wagmi": "^5.1.11",
+    "@zerodevx/svelte-toast": "^0.9.5",
     "abitype": "^1.0.2",
     "ajv": "^8.16.0",
     "autoprefixer": "^10.4.19",
+    "buffer": "^6.0.3",
     "daisyui": "^4.12.2",
+    "dayjs": "^1.11.13",
+    "debug": "^4.3.5",
     "dotenv": "^16.4.5",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-simple-import-sort": "^12.1.0",
     "eslint-plugin-svelte": "^2.35.1",
     "ethereum-address": "^0.0.4",
+    "events": "^3.3.0",
     "graphql": "^16.9.0",
+    "graphql-codegen": "^0.4.0",
     "graphql-codegen-svelte-apollo": "^1.1.0",
+    "graphql-tag": "^2.12.6",
     "jsdom": "^24.1.0",
+    "lefthook": "^1.7.12",
     "lokijs": "^1.5.12",
+    "node-conditions": "^1.2.0",
+    "object-hash": "^3.0.0",
     "postcss": "^8.4.38",
+    "postcss-import": "^16.1.0",
     "prettier": "^3.3.1",
     "prettier-eslint": "^16.3.0",
     "prettier-plugin-svelte": "^3.2.4",
+    "stores": "link:@zerodevx/svelte-toast/stores",
     "super-sitemap": "^1.0.3",
     "svelte": "^4.2.18",
+    "svelte-apollo": "^0.5.0",
     "svelte-check": "^3.8.0",
     "svelte-confetti": "^1.4.0",
+    "svelte-countdown": "^1.1.2",
+    "svelte-flip": "^0.0.11",
+    "svelte-i18n": "^4.0.0",
     "tailwindcss": "^3.4.4",
+    "trailblazer-contracts": "git+https://github.com:taikoxyz/trailblazer-contracts.git",
     "ts-morph": "^19.0.0",
     "tslib": "^2.6.3",
     "typescript": "^5.4.5",
+    "viem": "^2.21.57",
     "vite": "^5.2.13",
     "vite-plugin-qrcode": "^0.2.3",
     "vite-tsconfig-paths": "^4.3.2",
     "vitest": "^1.6.0",
     "vitest-fetch-mock": "^0.2.2",
     "vitest-mock-extended": "1.3.1",
-    "svelte-apollo": "^0.5.0",
-    "trailblazer-contracts": "git+https://github.com:taikoxyz/trailblazer-contracts.git",
-    "@sentry/sveltekit": "^8",
-    "graphql-codegen": "^0.4.0",
-    "graphql-tag": "^2.12.6",
-    "viem": "^2.21.57",
-    "@ethereumjs/util": "^9.0.3",
-    "@metamask/eth-sig-util": "^7.0.2",
-    "@sveltejs/adapter-node": "^5.2.5",
-    "@sveltejs/adapter-vercel": "^5.3.1",
-    "@sveltejs/vite-plugin-svelte": "^3.1.1",
-    "@wagmi/connectors": "^5.0.9",
-    "@wagmi/core": "^2.16.1",
-    "@walletconnect/core": "^2.17.3",
-    "@walletconnect/utils": "^2.17.3",
-    "@zerodevx/svelte-toast": "^0.9.5",
-    "buffer": "^6.0.3",
-    "dayjs": "^1.11.13",
-    "debug": "^4.3.5",
-    "events": "^3.3.0",
-    "lefthook": "^1.7.12",
-    "node-conditions": "^1.2.0",
-    "object-hash": "^3.0.0",
-    "postcss-import": "^16.1.0",
-    "stores": "link:@zerodevx/svelte-toast/stores",
-    "svelte-countdown": "^1.1.2",
-    "svelte-flip": "^0.0.11",
-    "svelte-i18n": "^4.0.0",
-    "wagmi": "^2.14.4",
-    "@walletconnect/web3wallet": "^1.16.1",
-    "@web3modal/wagmi": "^5.1.11"
+    "wagmi": "^2.14.4"
   }
 }
