+++ src/lib/shared/wagmi/client.ts
@@ -1,23 +1,17 @@
+import { WagmiAdapter } from '@reown/appkit-adapter-wagmi';
 import { cookieStorage, createStorage, reconnect } from '@wagmi/core';
-import { defaultWagmiConfig } from '@web3modal/wagmi';
 
 import { PUBLIC_WALLETCONNECT_PROJECT_ID } from '$env/static/public';
 import { chains } from '$lib/shared/utils/chain';
 
-const metadata = {
-  name: 'Taiko Trailblazer',
-  description: 'Taiko Trailblazer',
-  url: 'https://trailblazers.taiko.xyz/',
-  icons: ['https://avatars.githubusercontent.com/u/99078433'],
-};
-
-export const wagmiConfig = defaultWagmiConfig({
-  chains,
-  metadata,
+export const wagmiAdapter = new WagmiAdapter({
   projectId: PUBLIC_WALLETCONNECT_PROJECT_ID,
+  networks: chains,
   storage: createStorage({
     storage: cookieStorage,
   }),
 });
+
+export const wagmiConfig = wagmiAdapter.wagmiConfig;
 
 reconnect(wagmiConfig);
